refactor: separate script profile into
Separate script profile into profile manager for create or edit profile

and profile mode for mode (scroll view content)
