+++ keytik/profile_manager/__init__.py
@@ -1,0 +1,1 @@
+"""Profile manager package."""
