+++ keytik/profile_mode/__init__.py
@@ -1,0 +1,1 @@
+"""Profile mode packages."""
