--- keytik/script_profile/__init__.py
@@ -1,1 +1,0 @@
-"""Script profile package."""
